[ACS-8249] rename ui-header to header
