+++ docs/core/components/adf-header.component.md
@@ -1,0 +1,1 @@
+../../../lib/core/src/lib/header/header.component.md
