--- docs/core/components/ui-header.component.md
@@ -1,1 +1,0 @@
-../../../lib/core/src/lib/ui-header/ui-header.component.md
