+++ lib/core/src/lib/header/docs-images/adf-header-extended.png
