+++ lib/core/src/lib/header/docs-images/adf-header-minimal.png
