+++ lib/core/src/lib/header/header.component.scss
@@ -1,4 +1,4 @@
-.adf-ui-header {
+.adf-header {
     display: flex;
     flex-direction: column;
     width: var(--adf-header-width, 100%);
