+++ lib/core/src/lib/header/navbar/navbar-item.component.scss
