+++ lib/core/src/lib/header/navbar/navbar.component.html
@@ -2,4 +2,4 @@
                  [routerLink]="item.routerLink"
                  [label]="item.label">
 </adf-navbar-item>
-<ng-content ></ng-content>
+<ng-content></ng-content>
