+++ lib/core/src/lib/header/navbar/navbar.component.scss
